+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 17.2.0 (upcoming)
 
-* Pending changelog
+**New features:**
+
+* st-fullscreen-layout: Allow to display it on full width
+
 
 ## 17.1.0 (October 18, 2019)